fix where clause checks, add JSON / CSV input method, add test files

+++ run.sh
@@ -1,9 +1,19 @@
 #!/bin/bash
-rm -rf tmp*
-mvn clean package
-rm -rf ~/s4/build/s4-image/s4-apps/s4-latin/
-mkdir ~/s4/build/s4-image/s4-apps/s4-latin/
-cp -R *.s4latin ~/s4/build/s4-image/s4-apps/s4-latin/
-cp -R target/dist/ ~/s4/build/s4-image/s4-apps/s4-latin/
-sh start_app.sh 
+if [ -d "$S4_IMAGE" ]; then
+	echo "---------------------------------------------------"
+	echo "S4_IMAGE: $S4_IMAGE"
+	echo "---------------------------------------------------"
+	echo ""
+	rm -rf tmp*
+	mvn clean package
+	rm -rf $S4_IMAGE/s4-apps/s4-latin
+	mkdir $S4_IMAGE/s4-apps/s4-latin/
+	cp -R src/main/resources/*.s4latin $S4_IMAGE/s4-apps/s4-latin/
+	cp -R target/dist/ $S4_IMAGE/s4-apps/s4-latin/
+	sh start_app.sh 
+else
+   echo "No S4 image found."
+   echo "You need to set the environment variable S4_IMAGE to an existing S4 Image!"
+fi
+
